rewrites components as necessary
any component that needs either a) access to the current user object or b) to make authenticated api requests (or both) must be a function component, so it can consume the useStore and useAuth0 hooks. A lot of our components seem to have still been class-based, so have rewritten as required. Other existing class-based components that don't need to consume those hooks will be left as is for now, and rewritten in a future PR